feat: add execution generation
Handle dagger do <cmd> with string arguments.
Split code into multiples files and function.
Add local import and test documentation.
