+++ sdk/nodejs/example/simple-entrypoint/package.json
@@ -8,7 +8,7 @@
     "start": "node --loader ts-node/esm ./main.mts"
   },
   "dependencies": {
-    "@dagger.io/dagger": "^0.6.1",
+    "@dagger.io/dagger": "../../",
     "@types/node": "^20.2.5",
     "ts-node": "^10.9.1",
     "typescript": "^5.0.4"
