+++ sdk/nodejs/example/simple-entrypoint/tsconfig.json
