+++ sdk/nodejs/index.ts
@@ -4,4 +4,4 @@ export { default } from "./api/client.gen.js"
 export { gql } from "graphql-tag"
 export { GraphQLClient } from "graphql-request"
 export { connect, ConnectOpts, CallbackFct } from "./connect.js"
-export { serveCommands } from "./server.js"
+export { serveCommands } from "./server/server.js"
